新增服务端耗时-客户端

+++ sentinel-cluster/sentinel-cluster-common-default/src/main/java/com/hellobikemap/csp/sentinel/cluster/response/data/FlowTokenResponseData.java
@@ -30,7 +30,7 @@ public class FlowTokenResponseData {
         return serverCostTime;
     }
 
-    public FlowTokenResponseData setServerCostTime(int serverCostTime) {
+    public FlowTokenResponseData setServerCostTime(long serverCostTime) {
         this.serverCostTime = serverCostTime;
         return this;
     }
